assigment/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# LangChain and AI
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.5.0

# Database
sqlalchemy>=2.0.0
sqlite-utils>=3.35

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
numpy>=1.26.0
pandas>=2.2.0

# Logging and observability
loguru>=0.7.0
prometheus-client>=0.19.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: OpenAI (if using API)
# openai>=1.10.0
